refactor(alerts): share snooze date resolution
Generated-By: PostHog Desktop
Task-Id: 0ecd7f8a-92b2-4927-9bf6-6c393fba5833
