+++ products/logs/frontend/components/LogsAlerting/logsAlertingLogic.ts
@@ -7,6 +7,7 @@ import { lemonToast } from '@posthog/lemon-ui'
 import { dayjs } from 'lib/dayjs'
 import { teamLogic } from 'scenes/teamLogic'
 
+import { resolveSnoozeUntil } from 'products/alerts/frontend/utils'
 import {
     logsAlertsDestroy,
     logsAlertsList,
@@ -21,32 +22,6 @@ import {
 import { alertFiltersForPreEnableCheck, dispatchPreEnableCheck, runPreEnableChecks } from './logsAlertUtils'
 
 const ALERT_POLL_INTERVAL_MS = 30_000
-
-export function resolveSnoozeUntil(value: string): string {
-    const relativeValue = value.match(/^\+(\d+)([mhdwMy])$/)
-    if (!relativeValue) {
-        return dayjs(value).endOf('day').toISOString()
-    }
-
-    const amount = Number(relativeValue[1])
-    const unit = relativeValue[2]
-    if (unit === 'm') {
-        return dayjs().add(amount, 'minute').toISOString()
-    }
-    if (unit === 'h') {
-        return dayjs().add(amount, 'hour').toISOString()
-    }
-    if (unit === 'd') {
-        return dayjs().add(amount, 'day').toISOString()
-    }
-    if (unit === 'w') {
-        return dayjs().add(amount, 'week').toISOString()
-    }
-    if (unit === 'M') {
-        return dayjs().add(amount, 'month').toISOString()
-    }
-    return dayjs().add(amount, 'year').toISOString()
-}
 
 // Generated by kea-typegen. Update if you're an agent, ignore if you're human.
 export interface logsAlertingLogicValues {
